improvement(tools): smarter nixpkgs.json updates
When using `update.sh` to do lockfile updates, using `git rev-parse
HEAD` means we need to have a priori knowledge of where the
`nixpkgs.json` file is in the tree so we can write to it. This means if
users of this code want to put it somewhere other than `./nix` in the
root directory, they have to update the `update.sh` itself.

Instead, make a different assumption: that `update.sh` will always exist
next to `nixpkgs.json`. Then, it's enough for the script to query its
own location and do the rest.

Signed-off-by: Austin Seipp <aseipp@pobox.com>